修正群信息中的角色字段查询逻辑, 现在会返回role
修复了群信息逻辑中角色字段的查询逻辑。现在从群成员模型中正确加载群角色，以解决群信息响应中角色字段不正确的问题。此外，还预加载群成员列表，以提高查询效率，并确保通过正确的方法检查用户是否在群中。

+++ im_group/group_models/group_member_model.go
@@ -8,6 +8,6 @@ type GroupMemberModel struct {
 	GroupModel      GroupModel `gorm:"foreignKey:GroupID" json:"-"`   //群
 	UserID          uint       `json:"userID"`                        //用户id
 	MemberNickname  string     `gorm:"size:32" json:"memberNickname"` //群昵称
-	Role            int        `json:"role"`                          //1 群主 2 管理员 3 普通成员
+	Role            int8       `json:"role"`                          //1 群主 2 管理员 3 普通成员
 	ProhibitionTime *int       `json:"prohibitionTime"`               // 禁言时间 单位分钟
 }
